Convert video hero to hero variant
